+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Include transform and transformExtent OL methods. #171
+- Add getCurrentViewExtentCoordinates instance function. #171
 
 ## [v2.1.0] - 2022-11-15